==> 1-prepare_split.ipynb <==
import pandas as pd
import gc

import feather

import os
os.listdir('../data/')

data_path = '../data/'
click_train=pd.read_csv(data_path+'clicks_train.csv.zip',chunksize=100000)

## 取前10w个数据吧
c = 0 
df_click_train=None
for item  in click_train:
    df_click_train=item
    break

df_click_train.clicked.sum()

df_click_train.shape

click_test = pd.read_csv(data_path+'clicks_test.csv.zip',chunksize=100000)

df_click_test=None
for item in click_test:
    df_click_test = item
    break

df_click_test.shape

df_click_test.head()

display_train=set(df_click_train.display_id)

display_test=set(df_click_test.display_id)

a = display_train.intersection(display_test)

ad_train=set(df_click_train.ad_id)
ad_test=set(df_click_test.ad_id)

df_events=pd.read_csv(data_path+'events.csv.zip',chunksize=100000)


import tqdm
c = 0

for item in tqdm.tqdm(df_events):
    if c==0:
        c+=1
        df_train_events=item
        df_test_events=item
    item1 = item.loc[item.display_id.isin(df_click_train.display_id),:]
    df_train_events = df_train_events.append(item1)
    item2 = item.loc[item.display_id.isin(df_click_test.display_id),:]
    df_test_events = df_test_events.append(item2)
    

df_total_events=pd.concat([df_train_events,df_test_events],axis=0)

df_total_events.shape

def paths(tokens):
    all_paths = ['_'.join(tokens[0:(i+1)]) for i in range(len(tokens))]
    return ' '.join(all_paths)

def unwrap_geo(geo):
    geo = geo.split('>')
    return paths(geo)

df_total_events.geo_location.fillna('', inplace=True)

geo_str = df_total_events.geo_location.apply(unwrap_geo)


ts = (df_total_events.timestamp + 1465876799998) / 1000
df_total_events.timestamp = pd.to_datetime(ts, unit='s')

dt = df_total_events.timestamp.dt

dow = 'dow_' + dt.dayofweek.astype('str')
hours = 'hour_' + dt.hour.astype('str')
dow_hour = 'dow_hour_' + dt.dayofweek.astype('str') + '_' + dt.hour.astype('str')

display_str = 'u_' + df_total_events.uuid + ' ' + \
              'd_' + df_total_events.document_id.astype('str') + ' ' + \
              'p_' + df_total_events.platform.astype('str') + ' ' +  \
              dow + ' ' + hours + ' ' + dow_hour + ' ' + \
              geo_str

df_events_processed = pd.DataFrame()
df_events_processed['display_id'] = df_total_events.display_id
df_events_processed['display_str'] = display_str

df_promoted = pd.read_csv(data_path+"promoted_content.csv.zip")

ad_string = 'addoc_' + df_promoted.document_id.astype('str') + ' ' \
            'campaign_' + df_promoted.campaign_id.astype('str') + ' ' \
            'adv_' + df_promoted.advertiser_id.astype('str') 

df_promoted_processed = pd.DataFrame()
df_promoted_processed['ad_id'] = df_promoted.ad_id
df_promoted_processed['promoted_ad_str'] = ad_string

ad_to_idx = dict(zip(df_promoted_processed.ad_id, df_promoted_processed.index))
display_to_idx = dict(zip(df_events_processed.display_id,df_events_processed.index))

# processing data in batches

def prepare_batch(batch):
    batch = batch.reset_index(drop=1)
    
    promoted_idx = batch.ad_id.apply(ad_to_idx.get)
    promoted_ad_str = df_promoted_processed.promoted_ad_str.iloc[promoted_idx]

    ## 这行代码有问题吧？？？
    display_idx = batch.display_id.apply(display_to_idx.get)
#     display_str = df_events_processed.display_str.iloc[batch.display_id - 1]
    display_str = df_events_processed.display_str.loc[display_idx]
    
    promoted_ad_str.reset_index(drop=True, inplace=True)
    display_str.reset_index(drop=True, inplace=True)

    batch['ad_display_str'] = promoted_ad_str + ' ' + display_str
    return batch

def append_to_csv(batch, csv_file):
    props = dict(encoding='utf-8', index=False)
#     if not os.path.exists(csv_file):
    batch.to_csv(csv_file, **props)
#     else:
#         batch.to_csv(csv_file, mode='a', header=False, **props)

def delete_file_if_exists(filename):
    if os.path.exists(filename):
        os.remove(filename)

def chunk_dataframe(df, n):
#     for i in range(0, len(df), n):
    for i in range(0, 2000000,n):
        yield df.iloc[i:i+n]

## 每次展示都有一个唯一的display_id，所以display_id肯定不会重复


batch = prepare_batch(df_click_train)
batch['fold']=0
total_index=list(range(df_click_train.shape[0]))
import numpy as np
sam_ind = np.random.choice(total_index,int(len(total_index)/2),replace=False)
batch.loc[batch.iloc[sam_ind,].index,'fold']=1

append_to_csv(batch, './processed/svm_features_train.csv')

batch = prepare_batch(df_click_test)
append_to_csv(batch, './processed/svm_features_test.csv')
