==> 2-ftrl.ipynb <==
from math import exp, log, sqrt

# implementation taken from kaggle scripts:
# https://www.kaggle.com/sudalairajkumar/outbrain-click-prediction/ftrl-starter-with-leakage-vars/code


def hash_element(el, D):
    h = hash(el) % D
    if h < 0:
        h = h + D
    return h

def hash_elements(elements, D):
    return [hash_element(el, D) for el in elements]


class FtrlProximal(object):
    ''' Our main algorithm: Follow the regularized leader - proximal
        In short,
        this is an adaptive-learning-rate sparse logistic-regression with
        efficient L1-L2-regularization
        Reference:
        http://www.eecs.tufts.edu/~dsculley/papers/ad-click-prediction.pdf
    '''

    def __init__(self, alpha, beta, L1, L2, D, interactions):
        # parameters
        self.alpha = alpha
        self.beta = beta
        self.L1 = L1
        self.L2 = L2

        # feature related parameters
        self.D = D

        self.interactions = interactions

        # model
        # n: squared sum of past gradients
        # z: weights
        # w: lazy weights
        self.n = [0.0] * (D + 1)
        self.z = [0.0] * (D + 1)
        self.w = {}

    def to_indices(self, x):
        res = hash_elements(x, self.D)

        if self.interactions:
            sorted_x = sorted(x)
            len_x = len(sorted_x)

            for i in range(len_x):
                for j in range(i + 1, len_x):
                    h = hash_element(sorted_x[i] + '_' + sorted_x[j], self.D)
                    res.append(h)

        return res

    def predict(self, x):
        x_hashed = self.to_indices(x)
        return self.predict_hashed(x_hashed)

    def predict_hashed(self, x):
        ''' Get probability estimation on x
            INPUT:
                x: features
            OUTPUT:
                probability of p(y = 1 | x; w)
        '''

        # parameters
        alpha = self.alpha
        beta = self.beta
        L1 = self.L1
        L2 = self.L2

        # model
        n = self.n
        z = self.z
        w = {}

        # wTx is the inner product of w and x
        wTx = 0.

        indices = [0]
        for i in x:
            indices.append(i + 1)

        for i in indices:
            sign = -1. if z[i] < 0 else 1.  # get sign of z[i]

            # build w on the fly using z and n, hence the name - lazy weights
            # we are doing this at prediction instead of update time is because
            # this allows us for not storing the complete w
            if sign * z[i] <= L1:
                # w[i] vanishes due to L1 regularization
                w[i] = 0.0
            else:
                # apply prediction time L1, L2 regularization to z and get w
                w[i] = (sign * L1 - z[i]) / ((beta + sqrt(n[i])) / alpha + L2)

            wTx += w[i]

        # cache the current w for update stage
        self.w = w

        # bounded sigmoid function, this is the probability estimation
        return 1.0 / (1.0 + exp(-max(min(wTx, 35.0), -35.0)))

    def update(self, x, p, y):
        ''' Update model using x, p, y
            INPUT:
                x: a list of indices
                p: probability prediction of our model
                y: answer
            MODIFIES:
                self.n: increase by squared gradient
                self.z: weights
        '''

        # parameter
        alpha = self.alpha

        # model
        n = self.n
        z = self.z
        w = self.w

        # gradient under logloss
        g = p - y

        indices = [0]
        for i in x:
            indices.append(i + 1)

        # update z and n
        for i in indices:
            sigma = (sqrt(n[i] + g * g) - sqrt(n[i])) / alpha
            z[i] += g - sigma * w[i]
            n[i] += g * g

    def fit(self, x, y):
        x_hashed = self.to_indices(x)
        p = self.predict_hashed(x_hashed)
        self.update(x_hashed, p, y)

# implementation of auc is taken from ml_metrics:
# https://github.com/benhamner/Metrics/blob/master/Python/ml_metrics/auc.py

def tied_rank(x):
    """
    Computes the tied rank of elements in x.
    This function computes the tied rank of elements in x.
    Parameters
    ----------
    x : list of numbers, numpy array
    Returns
    -------
    score : list of numbers
            The tied rank f each element in x
    """
    sorted_x = sorted(zip(x,range(len(x))))
    r = [0 for k in x]
    cur_val = sorted_x[0][0]
    last_rank = 0
    for i in range(len(sorted_x)):
        if cur_val != sorted_x[i][0]:
            cur_val = sorted_x[i][0]
            for j in range(last_rank, i): 
                r[sorted_x[j][1]] = float(last_rank+1+i)/2.0
            last_rank = i
        if i==len(sorted_x)-1:
            for j in range(last_rank, i+1): 
                r[sorted_x[j][1]] = float(last_rank+i+2)/2.0
    return r

def auc(actual, posterior):
    """
    Computes the area under the receiver-operater characteristic (AUC)
    This function computes the AUC error metric for binary classification.
    Parameters
    ----------
    actual : list of binary numbers, numpy array
             The ground truth value
    posterior : same type as actual
                Defines a ranking on the binary numbers, from most likely to
                be positive to least likely to be positive.
    Returns
    -------
    score : double
            The mean squared error between actual and posterior
    """
    r = tied_rank(posterior)
    num_positive = len([0 for x in actual if x==1])
    num_negative = len(actual)-num_positive
    sum_positive = sum([r[i] for i in range(len(r)) if actual[i]==1])
    auc = ((sum_positive - num_positive*(num_positive+1)/2.0) /
           (num_negative*num_positive))
    return auc

import re
from time import time
from csv import DictReader
from time import time



spaces = re.compile(r' +')


# model parameters

alpha = 0.1
beta = 0.0
L1 = 2.0
L2 = 0.0

D = 2 ** 25

interactions = True
n_epochs = 1
show_auc = False

models = {}
models['0'] = FtrlProximal(alpha, beta, L1, L2, D, interactions)
models['1'] = FtrlProximal(alpha, beta, L1, L2, D, interactions)
model_full  = FtrlProximal(alpha, beta, L1, L2, D, interactions)

# training the models


t0 = time()

print('trainning models...')

for i in range(n_epochs):
    print('epoch %d...' % i)

    with open('./processed/svm_features_train.csv', 'r') as f:
        reader = DictReader(f)

        cnt = 0
        for row in reader:
            y = int(row['clicked'])

            x = spaces.split(row['ad_display_str'].strip())

            if row['fold'] == '0':
                fold = '1'
            else: # '1'
                fold = '0'

            models[fold].fit(x, y)
            model_full.fit(x, y)

            cnt = cnt + 1
            if cnt % 1000000 == 0:
                print('processed %dth row' % cnt)


print('training took %0.3fm' % ((time() - t0) / 60))


# validation and oof prediction

print('validating models...')

t0 = time()

all_y = {'0': [], '1': []}
all_pred = {'0': [], '1': []}

f_pred = {}
f_pred['0'] = open('ftrl_pred_0.txt', 'w')
f_pred['0'].write('y_actual,y_pred\n')

f_pred['1'] = open('ftrl_pred_1.txt', 'w')
f_pred['1'].write('y_actual,y_pred\n')

with open('./processed/svm_features_train.csv', 'r') as f:
    reader = DictReader(f)

    cnt = 0
    for row in reader:
        y = int(row['clicked'])
        fold = row['fold']

        x = spaces.split(row['ad_display_str'].strip())
        y_pred = models[fold].predict(x)

        all_y[fold].append(y)
        all_pred[fold].append(y_pred)
        f_pred[fold].write('%s,%s\n' % (y, y_pred))

        cnt = cnt + 1
        if cnt % 1000000 == 0:
            print('processed %dth row' % cnt)
        if show_auc and cnt % 5000000 == 0:
            auc0 = auc(all_y['0'], all_pred['0'])
            auc1 = auc(all_y['1'], all_pred['1'])
            print('auc: %.4f, %.4f' % (auc0, auc1))

auc0 = auc(all_y['0'], all_pred['0'])
auc1 = auc(all_y['1'], all_pred['1'])            
print('final auc: %.4f, %.4f' % (auc0, auc1))

f_pred['0'].close()
f_pred['1'].close()    

print('predict took %0.3fm' % ((time() - t0) / 60))
del all_y, all_pred


# predicting the results on test

print('applying the model to the test data...')

t0 = time()

# f_pred = open('ftrl_pred_test.txt', 'w')
# f_pred.write('y_pred\n')

# with open('./processed/svm_features_test.csv', 'r') as f:
#     reader = DictReader(f)

#     cnt = 0
#     for row in reader:
#         x = spaces.split(row['ad_display_str'].strip())
#         y_pred = model_full.predict(x)
#         f_pred.write('%s\n' % y_pred)

#         cnt = cnt + 1
#         if cnt % 1000000 == 0:
#             print('processed %dth row' % cnt)

# f_pred.close()

# print('predict took %0.3fm' % ((time() - t0) / 60))
