==> 6_1_generate_ffm_data.ipynb <==
#_*_coding:utf-8_*_
import os
import pandas as pd
import numpy as np

df_train = pd.read_csv('./processed/svm_features_train.csv')

df_train.head()

D = 2 ** 20

# display features
USER = '0'
ON_DOC = '1'
PLATFORM = '2'

# ads features
AD = '3'
AD_DOC = '4'
CAMPAIGN = '5'
ADVERTISER = '6'

# document features
ON_SRC = '7'
ON_PUBLISHER = '8'

AD_SRC = '9'
AD_PUBLISHER = '10' 


def hash_element(el):
    h = hash(el) % D
    if h < 0:
        h = h + D
    return str(h)

df_events = pd.read_csv('../data/events.csv.zip',chunksize=1000000)

c = 0
train_events=None
for item in df_events:
    c+=1
    tar_item = item.loc[item.display_id.isin(df_train.display_id),:]
    
    if train_events is None:
        train_events=tar_item
    else:
        tar_shape=tar_item.shape[0]
        train_events_shape=train_events.shape[0]
        train_events.append(tar_item)
        # 确保 append 方法使用正确
        assert train_events.shape[0]== tar_shape+train_events_shape

del df_events

user_str = USER + ':' + train_events.uuid.apply(hash_element)+':1'
doc_str = ON_DOC + ':' + train_events.document_id.apply(hash_element)+':1'
platforms = PLATFORM +':' + train_events.platform.astype('str')+':1'

df_events_processed = pd.DataFrame()
df_events_processed['display_str'] = user_str + ' ' + doc_str + ' ' + platforms
df_events_processed['document_id'] = train_events.document_id
del user_str, doc_str, platforms

df_events_processed.shape

df_events_processed.head()

df_ads = pd.read_csv("../data/promoted_content.csv.zip")

ad_to_doc = dict(zip(df_ads.ad_id, df_ads.document_id))

ad_str = AD + ':' + df_ads.ad_id.astype(str) + ':1 ' + \
         AD_DOC + ':' + df_ads.document_id.apply(hash_element) + ':1 ' + \
         CAMPAIGN + ':' + df_ads.campaign_id.astype(str) + ':1 ' + \
         ADVERTISER + ':' + df_ads.advertiser_id.astype(str) + ':1'

ad_str_dict = dict(zip(df_ads.ad_id, ad_str))

del ad_str, df_ads

len(ad_str_dict)

df_train.head()

##以batch 的方式替换新建数据
## 根据display_id  和 ad_id 新建样本

display_ids = df_train.display_id
ad_ids = df_train.ad_id

display_str = df_events_processed.loc[display_ids-1,:]['display_str'].values

display_str.shape[0]

ad_str = [ad_str_dict[item] for item in ad_ids.values]

labels=df_train.clicked

## 遍历并保存到文件中
c = 0
with open('./processed/ffm_data.txt','w') as fh:
    for i in range(display_str.shape[0]):
        line = str(labels[0]) +' '+ display_str[i] + ' '+ad_str[i]
        fh.write(line)
        fh.write('\n')
        
        
    
    

## 转化成 deepfm 的格式
## batch_size * field_size 

df_deepfm = pd.read_csv('./processed/ffm_data.txt',sep=' ',header=None)


df_deepfm.head()

field_size=len(df_deepfm.columns)-1

def get_feature_index(row):
    items = row.split(':')
    return items[1]

def get_feature_values(row):
    items = row.split(':')
    return items[2]

indexs=[]
values=[]
for i in range(1,field_size+1):
    index_values = df_deepfm[i].apply(get_feature_index)
    indexs.append(index_values)
    
    values_values = df_deepfm[i].apply(get_feature_values)
    values.append(values_values)
    

indexs=np.array(indexs)
values = np.array(values)

indexs = np.transpose(indexs)
values=np.transpose(values)

labels=df_deepfm[0]

training_deepfm = {
    'xi':indexs,
    'vi':values,
    'label':labels
}

import  pickle
with open('./processed/deepfm_training.pickle','wb') as fh:
    pickle.dump(training_deepfm,fh)

pd.__version__
